refactor(analysis): extract Rust function complexity owner
